gpu: nvgpu: fix MISRA errors in runlist
Fix some mistakes from commit 0fbc1a265240 (gpu: nvgpu: avoid recursion in
runlist construction) and commit 998bf379df11 (gpu: nvgpu: add
runlist_append_tsg) for MISRA rules 10.3 and 10.4.

- cast a sizeof to u32 in a calculation to match in size,
- make the NVGPU_FIFO_RUNLIST_INTERLEAVE_LEVEL_* constants unsigned to make
  comparisons match in signedness.

Jira NVGPU-1174

Reviewed-on: https://git-master.nvidia.com/r/1938069
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>

+++ drivers/gpu/nvgpu/gk20a/fifo_gk20a.c
@@ -3330,7 +3330,7 @@ static u32 nvgpu_runlist_append_tsg(struct gk20a *g,
 		struct tsg_gk20a *tsg)
 {
 	struct fifo_gk20a *f = &g->fifo;
-	u32 runlist_entry_words = f->runlist_entry_size / sizeof(u32);
+	u32 runlist_entry_words = f->runlist_entry_size / (u32)sizeof(u32);
 	struct channel_gk20a *ch;
 	u32 count = 0;
